move LocaleSelector component
